feat(rules): adds requirements.Store with requirements context
This change adds support for processing OSCAL Control Implementations
and using that processed data to apply settings and context from requirements to
RuleSet results in a rules.Store implementation. The requirements.Store implementation
intentionally does not change the indexed RuleSets in the base rules.Store to avoid
unexpected behavior.

The All() method is also removed from the rules.Store interface for simplification.

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>